backend/requirements.txt
fastapi
uvicorn[standard]
transformers
torch
easyocr
Pillow
numpy
python-multipart
httpx
beautifulsoup4
